chore: 인과관련 memberId를 String에서 Long으로 변경

+++ doonut-app-external-api/src/main/kotlin/com/doonutmate/oauth/configuration/AuthorizationArgumentResolver.kt
@@ -36,7 +36,7 @@ class AuthorizationArgumentResolver(
         }
 
         val token = authorizationHeader.substring(BEARER_PREFIX_LEN)
-        return jwtTokenProvider.getPayload(token)
+        return jwtTokenProvider.getPayload(token).toLong()
     }
 
     companion object {
